mini_project_3/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.17.0
openai==1.109.1
openai-agents==0.3.3
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
youtube-transcript-api==1.2.2
